chore: Continued refactor of appdata container logic
